+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/MockClientTest.java
@@ -447,4 +447,5 @@ public class MockClientTest {
 		client.updateStepResult(stepResult);
 		client.deleteStepResult(0L);
 	}
+	
 }